pyproject.toml
[project]
name = "cmm-backend"
version = "0.1.0"
description = "CMM Backend — FastAPI + Supabase (PostgreSQL)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "supabase>=2.10.0",
    "boto3>=1.35.0",
    "pydantic-settings>=2.0.0",
    "pyairtable>=3.3.0",
    "sqlalchemy[postgresql]>=2.0.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
scripts = [
    "pandas>=2.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
